下ドロワーの各機能をコンポーネントに切り出し
