+++ app/javascript/react/canvas/components/fetch_city_data/bottom_drawer.jsx
@@ -1,10 +1,10 @@
 import React, { useState } from 'react';
 import Box from '@mui/material/Box';
-// import SwipeableDrawer from '@mui/material/SwipeableDrawer';
 import Drawer from '@mui/material/Drawer';
 import ClickAwayListener from '@mui/material/ClickAwayListener';
 
-import GoogleMapComponent from './GoogleMapComponent';
+import CitySearchBox from './city_search_box';
+import GoogleMapComponent from './googlemap_component';
 
 export default function BottomDrawer({open, handleClose, bottomDrawerButtonRef, setCityId}) {
   // ドロワーの状態，ドロワー閉ハンドラ，下ドロワー起動ボタンのref，setCityId関数
@@ -18,20 +18,6 @@ export default function BottomDrawer({open, handleClose, bottomDrawerButtonRef, 
     handleClose();
   }
   
-  // セレクトボックスで選択中の都市情報を保持するステート
-  const [selectedCity, setSelectedCity] = useState(null);
-  
-  // 都市名セレクトボックスの変更ハンドラ
-  const handleChange = (e, value) => {
-    setSelectedCity(value);
-  }
-  // 都市名選択の反映実行ボタンクリックのハンドラ
-  const handleButtonClick = () => {
-    if (selectedCity) {
-      setCityId(selectedCity.id);   //index.jsxから引き受けたsetCityId関数を実行し，選択した都市IDを更新
-    }
-  }
-
   // モックデータ
   const cityIdMapping = [
     { id: 1, name: '東京', position: { lat: 35.6917, lng: 139.75, alt: 25.2 }},
@@ -49,34 +35,7 @@ export default function BottomDrawer({open, handleClose, bottomDrawerButtonRef, 
         // onClose={handleClose}
       >
         {/* カラムから呼び出し */}
-        {/* <Box 
-          sx={{ 
-            width: '100%',
-            // height: '200px',
-            margin: '30px',
-            display: 'flex',
-            flexDirection: 'row',
-            justifyContent: 'center',
-            alignItems: 'flex-start',
-          }}
-          role="presentation">
-          <Autocomplete
-          onChange={handleChange}
-          disablePortal
-          options={cityIdMapping}
-          getOptionLabel={(option) => option.name}
-          sx={{ width: 300, height: 50}}
-          renderInput={(params) => <TextField {...params} label="都市を選択" />}
-          />
-          <Button 
-            onClick={handleButtonClick}
-            variant='contained'
-            size='large'
-            sx={{ height: 50, marginLeft: '20px' }}
-          >
-            グラフに反映
-          </Button>
-        </Box> */}
+        <CitySearchBox cityIdMapping={cityIdMapping} setCityId={setCityId}/>
         
         {/* GoogleMapから呼び出し */}
         <Box
