+++ app/javascript/react/canvas/components/fetch_city_data/googlemap_component.jsx
